refactor: TextApplicationController 분리
- 기존 ApplicationController은 ConsoleApplicationController로 이름 변경
